Fix Gson type mismatches aborting image + query rows
The image/domain records carry index:0.0 and the Queries[] carry count:46.0
(JSON doubles), but ImageRec.index was Integer and Query.count was Long, so
gson.fromJson(record, ...) threw NumberFormatException inside emitImages -> the
outer catch swallowed it and skipped Thumbnail/Aligned-to/Downloads AND the
Query-for menu (browser diff vs prod confirmed exactly these four rows missing).
Make both fields Double (index unused; count via longValue()).

+++ src/main/java/uk/ac/vfb/geppetto/VFBProcessTermInfoVFBqueryJson.java
@@ -169,7 +169,7 @@ public class VFBProcessTermInfoVFBqueryJson extends AQueryProcessor {
 		String obj;
 		String wlz;
 		String swc;
-		Integer index;
+		Double index;
 		String orientation;
 		XYZ center;
 		XYZ extent;
@@ -192,7 +192,7 @@ public class VFBProcessTermInfoVFBqueryJson extends AQueryProcessor {
 	private class Query {
 		String query;            // query type id, e.g. "SplitsTargeting"
 		String label;            // display label
-		Long count;              // result count (may be null / -1 for deferred)
+		Double count;            // result count (JSON sends 46.0; null/-1 = deferred)
 	}
 
 	// ---- main process -------------------------------------------------------
@@ -734,7 +734,7 @@ public class VFBProcessTermInfoVFBqueryJson extends AQueryProcessor {
 		for (JsonElement el : ti.getAsJsonArray("Queries")) {
 			Query q = g.fromJson(el, Query.class);
 			if (q == null || q.query == null) continue;
-			long count = q.count == null ? -1 : q.count;
+			long count = q.count == null ? -1 : q.count.longValue();
 			String badge;
 			String cssExtra = "";
 			if (count == 0) {
